% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm-functions.r
\name{bestSequence}
\alias{bestSequence}
\title{Given a position weight matrix, find the best matching sequence}
\usage{
bestSequence(pwm)
}
\arguments{
\item{pwm}{Position weight matrix}
}
\description{
Finds the amino acid at each position of the PWM with the highest occurence.
Used in matrix similarity score calculation.
}
\examples{
# No Examples
}
\keyword{best}
\keyword{internal}
\keyword{pwm}

