% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimp.R
\name{computeBinding}
\alias{computeBinding}
\title{Score wt and mt sequences for a pwm}
\usage{
computeBinding(obj, mut_ss, mut_location, prob.thresh = 0.5,
  log2.thresh = 1)
}
\arguments{
\item{obj}{MIMP object containing PWM, GMM parameters, and etc.}

\item{mut_ss}{snvs data frame containing wt and mt sequences computed from SNVs function}

\item{mut_location}{list of mutation locations}

\item{prob.thresh}{Probability threshold of gains and losses. This value should be between 0.5 and 1.}

\item{log2.thresh}{Threshold for the absolute value of log ratio between wild type and mutant scores. Anything less than this value is discarded (default: 1).}
}
\description{
Score wt and mt sequences for a pwm
}
\keyword{mut}
\keyword{score}
\keyword{snp}
\keyword{snv}
