% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.r
\name{dohtmlSh3}
\alias{dohtmlSh3}
\title{Helper function for \code{results2html} and sh3 binding}
\usage{
dohtmlSh3(x, LOGO_DIR, HL_DIR, logoExt = ".svg", .webserver = F)
}
\arguments{
\item{x}{Data frame resulting from mimp call.}

\item{LOGO_DIR}{Directory containing sequence logo images.}

\item{HL_DIR}{Directory containing overlays}

\item{logoExt}{Extension of logo files}

\item{.webserver}{Request coming from webserver?}
}
\description{
Helper function for \code{results2html} and sh3 binding
}
\keyword{display}
\keyword{mimp}
