\name{extractDigits}
\alias{extractDigits}
\title{Extracts digits from a string and returns them in a numerical form}
\usage{
extractDigits(string)
}
\arguments{
  \item{string}{String to be manipulated}
}
\description{
Extracts digits from a string and returns them in a
numerical form
}
\examples{
extractDigits('A123F')
}
\keyword{digits}
\keyword{numerical}

