% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimp.R
\name{mimp}
\alias{mimp}
\title{Predict the impact of single variants on phosphorylation.}
\usage{
mimp(muts, seqs, psites = NULL, prob.thresh = 0.5, log2.thresh = 1,
  display.results = T, include.cent = F, model.data = "hconf")
}
\arguments{
\item{muts}{Mutation data file: a space delimited text file OR data frame containing two columns (1) gene and (1) mutation. 
Example:
\tabular{ll}{
   TP53 \tab R282W\cr
   CTNNB1 \tab S33C\cr
   CTNNB1 \tab S37F\cr
}}

\item{seqs}{Sequence data file containing protein sequences in FASTA format OR 
named list of sequences where each list element is the uppercase sequence and the 
name of each element is that of the protein. Example: list(GENEA="ARNDGH", GENEB="YVRRHS")}

\item{psites}{Phosphorylation data file (optional): a space delimited text file OR data frame containing  two columns (1) gene and (1) positions of phosphorylation sites. Example:
\tabular{ll}{
   TP53 \tab 280\cr
   CTNNB1 \tab 29\cr
   CTNNB1 \tab 44\cr
}}

\item{prob.thresh}{Probability threshold of gains and losses. This value should be between 0.5 and 1.}

\item{log2.thresh}{Threshold for the absolute value of log ratio between wild type and mutant scores. Anything less than this value is discarded (default: 1).}

\item{include.cent}{If TRUE, gains and losses caused by mutation in the central STY residue are kept. Scores of peptides with a non-STY central residue is given a score of 0 (default: FALSE).}

\item{model.data}{Name of specificity model data to use, can be "hconf" : individual experimental kinase specificity models used to scan for rewiring events. For experimental kinase specificity models, grouped by family, set to "hconf-fam". Both are considered high confidence. For lower confidence predicted specificity models , set to "lconf". NOTE: Predicted models are purely speculative and should be used with caution}
}
\value{
The data is returned in a \code{data.frame} with the following columns:
\item{gene}{Gene with the rewiring event}
\item{mut}{Mutation causing the rewiring event}
\item{psite_pos}{Position of the phosphosite}
\item{mut_dist}{Distance of the mutation relative to the central residue}
\item{wt}{Sequence of the wildtype phosphosite (before the mutation). Score is NA if the central residue is not S, T or Y}
\item{mt}{Sequence of the mutated phosphosite (after the mutation). Score is NA if the central residue is not S, T or Y}
\item{score_wt}{Matrix similarity score of the wildtype phosphosite}
\item{score_mt}{Matrix similarity score of the mutated phosphosite}
\item{log_ratio}{Log2 ratio between mutant and wildtype scores. A high positive log ratio represents a high confidence gain-of-phosphorylation event. A high negative log ratio represents a high confidence loss-of-phosphorylation event. This ratio is NA for mutations that affect the central phosphorylation sites}
\item{pwm}{Name of the kinase being rewired}
\item{pwm_fam}{Family/subfamily of kinase being rewired. If a kinase subfamily is available the family and subfamily will be separated by an underscore e.g. "DMPK_ROCK". If no subfamily is available, only the family is shown e.g. "GSK"}
\item{nseqs}{Number of sequences used to construct the PWM. PWMs constructed with a higher number of sequences are generally considered of better quality.}
\item{prob}{Joint probability of wild type sequence belonging to the foreground distribution and mutated sequence belonging to the background distribution, for loss and vice versa for gain.}
\item{effect}{Type of rewiring event, can be "loss" or "gain"}
}
\description{
This function takes in mutation, sequence and phosphorylation data to predict the 
impact the mutation has on phosphorylation.
}
\examples{
# Get the path to example mutation data 
mut.file = system.file("extdata", "mutation_data.txt", package = "rmimp")

# Get the path to example FASTA sequence data 
seq.file = system.file("extdata", "sequence_data.txt", package = "rmimp")

# View the files in a text editor
browseURL(mut.file)
browseURL(seq.file)

# Run rewiring analysis
results = mimp(mut.file, seq.file, display.results=TRUE)

# Show head of results
head(results)
}
\keyword{kinase}
\keyword{mimp}
\keyword{mutation}
\keyword{phosphorylation}
\keyword{psites}
\keyword{pwm}
\keyword{rewiring}
\keyword{snp}
\keyword{snv}

