% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm-functions.r
\name{mss}
\alias{mss}
\title{Compute matrix similarity score as described in MATCH algorithm}
\usage{
mss(seqs, pwm, na_rm = F, ignore_cent = T, kinase.domain = T,
  residues_groups = c("S|T", "Y"))
}
\arguments{
\item{seqs}{Sequences to be scored}

\item{pwm}{Position weight matrix}

\item{na_rm}{Remove NA scores?}

\item{ignore_cent}{If TRUE, central residue is ignore from scoring.}

\item{kinase.domain}{Whether the domain to be trained is a kinase domain.}

\item{residues_groups}{a vector of regular expressions used to group kinases by central residue they target;
if a sequence does not have a central residue matching a group chosen from modified.residues by the algorithm
(based on PWM), the sequence will be discarded.}
}
\description{
Computes matrix similarity score of a PWM with a k-mer.
Score ranges from 0-1, as described in [PMID: 12824369]
}
\examples{
# No Examples
}
\keyword{internal}
\keyword{match}
\keyword{mss}
\keyword{pwm}
\keyword{tfbs}
