% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimp.R
\name{predictKinasePhosphosites}
\alias{predictKinasePhosphosites}
\title{Compute posterior probability of wild type phosphosites for kinases}
\usage{
predictKinasePhosphosites(psites, seqs, model.data = "hconf",
  posterior_thresh = 0.8, intermediate = F, kinases)
}
\arguments{
\item{psites}{phosphorylation data, see \code{?mimp} for details}

\item{seqs}{sequence data, see \code{?mimp} for details}

\item{model.data}{MIMP model used, see \code{?mimp} for details}

\item{posterior_thresh}{posterior probability threshold that the score belongs to the foreground distribution of the kinase, probabilities below this value are discarded (default 0.8)}

\item{intermediate}{if TRUE intermediate MSS scores and likelihoods are reported (default FALSE)}

\item{kinases}{vector of kinases used for the scoring (e.g. c("AURKB", "CDK2")), if this isn't provided all kinases will be used .}
}
\value{
The data is returned in a \code{data.frame} with the following columns:
\item{gene}{Gene with the rewiring event}
\item{pos}{Position of the phosphosite}
\item{wt}{Sequence of the wildtype phosphosite}
\item{score_wt}{(intermediate value) matrix similarity score of sequence}
\item{l.wt.fg}{(intermediate value) likelihood of score given foreground distribution}
\item{l.wt.bg}{(intermediate value) likelihood of score given background distribution}
\item{post.wt.fg}{posterior probability of score in foreground distribution}
\item{post.wt.bg}{posterior probability of score in background distribution}
\item{pwm}{Name of the predicted kinase}
\item{pwm_fam}{Family/subfamily of the predicted kinase. If a kinase subfamily is available the family and subfamily will be seprated by an underscore e.g. "DMPK_ROCK". If no subfamily is available, only the family is shown e.g. "GSK"}
}
\description{
Compute posterior probability of wild type phosphosites for kinases
}
\examples{
# Get the path to example phosphorylation data 
psite.file = system.file("extdata", "sample_phosphosites.tab", package = "rmimp")

# Get the path to example FASTA sequence data 
seq.file = system.file("extdata", "sample_seqs.fa", package = "rmimp")

# Run for all kinases
results_all = predictKinasePhosphosites(psite.file, seq.file)

# Run for select kinases
results_select = predictKinasePhosphosites(psite.file, seq.file, kinases=c("AURKB", "CDK2"))
}

