% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pRewiringPosterior}
\alias{pRewiringPosterior}
\title{Computing posterior probability - ploss and pgain}
\usage{
pRewiringPosterior(wt.score, mt.score, dist.params, auc, show.likelihood = F)
}
\arguments{
\item{wt.score}{Wild type score}

\item{mt.score}{Mutant score}

\item{dist.params}{Distribution parameters of GMMs (foreground and background). This is precomputed and comes built into mimp.}

\item{auc}{AUC of the model. This is precomputed and comes built into mimp.}

\item{show.likelihood}{If TRUE, individual liklihoods used to compute ploss and pgain is returned. Otherwise only ploss and pgain returned}
}
\description{
Computing posterior probability - ploss and pgain
}

