% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimp.R
\name{pSNVs}
\alias{pSNVs}
\title{Find phosphorylation related variants (pSNVs)}
\usage{
pSNVs(md, pd, seqdata, flank = 7)
}
\arguments{
\item{flank}{Number of amino acids flanking the psite to be considered}

\item{muts}{Mutation data as data frame of two columns (1) name of gene or protein
(2) mutation in the format X123Y, where X is the reference amino acid
and Y is the alternative amino acid.}

\item{psites}{Phosphorylation data as a data frame of two columns (1) name of gene
or protein (2) Position of the phosphorylated residue}

\item{seqs}{Sequence data as a name list. Names of the list correspond to the gene or
protein name. Each entry contains the collapsed sequence.}
}
\description{
Given mutation data and psites, find variants that
exist in the flanking regions of the psite
}
\examples{
# No examples
}
\keyword{internal}
\keyword{mutation}
\keyword{psites}
\keyword{psnv}
\keyword{snp}
