% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm-functions.r
\name{PWM}
\alias{PWM}
\title{Construct position weight matrix}
\usage{
PWM(seqs, pseudocount = 0.01, relative.freq = T, is.kinase.pwm = T,
  priors = AA_PRIORS_HUMAN, do.pseudocounts = F)
}
\arguments{
\item{seqs}{Aligned sequences all of the same length}

\item{pseudocount}{Pseudocount factor. Final pseudocount is background probability * this factor}

\item{relative.freq}{Set to TRUE if each column should be divided by the sum}

\item{is.kinase.pwm}{Set to TRUE if matrix is being built for a kinase}

\item{priors}{Named character vector containing priors of amino acids.}

\item{do.pseudocounts}{TRUE if we are to add pseudocounts}
}
\description{
Makes a position weight matrix given aligned sequences.
}
\examples{
# No examples
}
\keyword{construct}
\keyword{internal}
\keyword{pwm}
