\name{replaceChar}
\alias{replaceChar}
\title{Replace charachters at certain positions of a string with another charachter.}
\usage{
replaceChar(string, pos, char)
}
\arguments{
  \item{string}{String to be manipulated}

  \item{pos}{One or more positions corresponding to
  charachters to be changed}

  \item{char}{Replacement charachter}
}
\description{
Replace charachters at certain positions of a string with
another charachter.
}
\examples{
replaceChar('ABC', 2, 'X')
}
\keyword{char}
\keyword{replace}

