% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.r
\name{results2html}
\alias{results2html}
\title{Display MIMP results interactively in browser}
\usage{
results2html(x, max.rows = 5000)
}
\arguments{
\item{x}{Data frame resulting from mimp call.}

\item{max.rows}{If data contains more rows than this value, results won't be displayed.}
}
\description{
Display MIMP results interactively in browser
}
\keyword{display}
\keyword{mimp}

