% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm-functions.r
\name{scoreArrayFast}
\alias{scoreArrayFast}
\title{Get weight/probability for each amino acid in a sequence}
\usage{
scoreArrayFast(seqs, pwm, do_sum = T, ignore_cent = F)
}
\arguments{
\item{seqs}{One or more sequences to be processed}

\item{pwm}{Position weight matrix}

\item{do_sum}{If TRUE sum position-based scores per sequence}

\item{ignore_cent}{If TRUE, ignore central residue before returning}
}
\description{
Gets weight/probability for the amino acid at each position of the sequence
as an array.
}
\examples{
# No Examples
}
\keyword{internal}
\keyword{match}
\keyword{mss}
\keyword{pwm}
\keyword{tfbs}

