% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwm-functions.r
\name{scoreArrayRolling}
\alias{scoreArrayRolling}
\title{Get weight/probability for each amino acid in a sequence}
\usage{
scoreArrayRolling(seqs, pwm)
}
\arguments{
\item{seqs}{One or more sequences to be processed}

\item{pwm}{Position weight matrix}
}
\description{
Gets weight/probability for the amino acid at each position of the sequence
as an array.
}
\examples{
# No Examples
}
\keyword{match}
\keyword{mss}
\keyword{pwm}
\keyword{tfbs}
