% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimp.R
\name{scoreWTSequence}
\alias{scoreWTSequence}
\title{Score wt sequence using PWMs in the model}
\usage{
scoreWTSequence(wt_seqs, central = T, domain = "phos", species = "human",
  model.data = "hconf", cores = 2)
}
\arguments{
\item{wt_seqs}{A list of sequences to be scored}

\item{central}{Whether the mutation site is at the central residue of the sequence}

\item{cores}{Number of cores the function could use}
}
\description{
Score wt sequence using PWMs in the model
}
