% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimp.R
\name{SNVs}
\alias{SNVs}
\title{Find non-central variants (SNVs)}
\usage{
SNVs(md, seqdata, flank)
}
\arguments{
\item{md}{Mutation data as data frame of two columns (1) name of gene or protein
(2) mutation in the format X123Y, where X is the reference amino acid
and Y is the alternative amino acid.}

\item{seqdata}{Phosphorylation data as a data frame of two columns (1) name of gene
or protein (2) Position of the phosphorylated residue}

\item{flank}{Number of amino acids flanking the site to be considered}
}
\description{
Given mutation data, find variants that
exist in the flanking regions of the psite
}
\examples{
# No examples
}
\keyword{mutation}
\keyword{snp}
\keyword{snv}
