% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmimp.R
\name{trainModel}
\alias{trainModel}
\title{Train GMM model
and return as a list to be used later. If file is passed,
the model will also be save to a .mimp file.}
\usage{
trainModel(pos.dir, neg.dir, kinase.domain = F, cores = 2, file = NULL,
  threshold = 10, min.auc = 0.65, priors)
}
\arguments{
\item{pos.dir}{the path to the directory contains positive entries}

\item{neg.dir}{the path to the directory contains negative entries}

\item{kinase.domain}{Whether the domain to be trained is a kinase domain.}

\item{cores}{(optional) the number of CPU cores that can be used to train the model}

\item{file}{(optional) the path to save the model}

\item{threshold}{(optional) the minimum number of scores needed for each domain to train the model}

\item{min.auc}{(optional) the minimum number of AUC needed for each domain to train the model}
}
\value{
a GMM model
}
\description{
Train GMM model
and return as a list to be used later. If file is passed,
the model will also be save to a .mimp file.
}
\examples{
No examples

}
