\name{worstSequence}
\alias{worstSequence}
\title{Given a position weight matrix, find the worst matching sequence}
\usage{
worstSequence(pwm)
}
\arguments{
  \item{pwm}{Position weight matrix}
}
\description{
Finds the amino acid at each position of the PWM with the
lowest occurence. Used in matrix similarity score
calculation.
}
\examples{
# No Examples
}
\keyword{pwm}
\keyword{worst}

