% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display-functions.r
\name{dohtml}
\alias{dohtml}
\title{Helper function for \code{results2html}}
\usage{
dohtml(x, LOGO_DIR, HL_DIR, .webserver = F)
}
\arguments{
\item{x}{Data frame resulting from mimp call.}

\item{LOGO_DIR}{Directory containing sequence logo images.}

\item{HL_DIR}{Directory containing overlays}

\item{.webserver}{Request coming from webserver?}
}
\description{
Helper function for \code{results2html}
}
\keyword{display}
\keyword{internal}
\keyword{mimp}

