\name{dohtml}
\alias{dohtml}
\title{Helper function for \code{results2html}}
\usage{
dohtml(x, LOGO_DIR)
}
\arguments{
  \item{x}{Data frame resulting from mimp call.}

  \item{LOGO_DIR}{Directory containing sequence logo
  images.}
}
\description{
Helper function for \code{results2html}
}
\keyword{display}
\keyword{mimp}

