% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PWM}
\alias{PWM}
\title{Construct position weight matrix}
\usage{
PWM(seqs, pseudocount = 0.01, relative.freq = T, type = "AA",
  priors = AA_PRIORS_HUMAN, log.bg = F)
}
\arguments{
\item{seqs}{Aligned sequences all of the same length}

\item{pseudocount}{Pseudocount factor. Final pseudocount is background probability * this factor}

\item{relative.freq}{TRUE if each column should be divided by the sum}

\item{type}{Type of sequences 'AA' or 'DNA'}

\item{priors}{Named character vector containing priors of amino acids.}

\item{log.bg}{If true, relative frequencies will be converted to weights using -log of freq/bgfreq}
}
\description{
Makes a position weight matrix given aligned sequences.
}
\examples{
# No examples
}
\keyword{construct}
\keyword{pwm}

