\name{unfactor}
\alias{unfactor}
\title{Converts all columns of a data frame of class factor to character}
\usage{
unfactor(df)
}
\arguments{
  \item{string}{String to be manipulated}
}
\description{
Converts all columns of a data frame of class factor to
character
}
\examples{
unfactor( data.frame(x=c('A', 'B')) )
}
\keyword{character}
\keyword{factor}

